Create blank.yml
i need someone to review this and edit it once reviewed. the edit should read "this is for those developers who wish to program for a purpose that will develop the human experience in itself. If you are wondering what that means, then just ask about it or ask directly. I am a developer, which means I develop ideas, I may not know how to program but I know how a program works. IF you are a first comer and would like to contribute, then know I mean no harm, I mean no complications to others. I only wish to build different apps and I will only establish this idea because of the covid-19 virus.  

TOMBSTONE. that's whats it's the outcome will be.... just Shhhhhhhhh, let us see how many avoid the read me files because the read me files are important. So do not ruin the master plan, just go with the flow. I will do my part if you do yours... One-piece at a time, similar to the great men of the early united states,  I will proceed without mystifying the process...

Viva the enlightened, neutral, and unbiased...... For the benefit of all, and the downfall of a few, the ones who could not see beyond their emotions, and their biases... If that is you, then consider this a cease and desist..... You know not with what you meddle.